Clean up Editor Configuration initialization
